+++ fuzz/fuzz_targets/json_differential.rs
@@ -17,13 +17,19 @@ static SETUP: Once = Once::new();
 
 fuzz_target!(|data: ArbitraryValue| {
     SETUP.call_once(|| {
+        let mut ref_config = Config::default();
+        setup_config(&mut ref_config);
+
         let mut config = Config::default();
+        setup_config(&mut config);
         config.simd_json_builtins(true).javy_json(true);
 
         unsafe {
             RT = Some(Runtime::new(std::mem::take(&mut config)).expect("Runtime to be created"));
-            REF_RT =
-                Some(Runtime::new(Config::default()).expect("Reference runtime to be created"));
+            REF_RT = Some(
+                Runtime::new(std::mem::take(&mut ref_config))
+                    .expect("Reference runtime to be created"),
+            );
         };
     });
 
@@ -36,9 +42,18 @@ fn exec(data: &ArbitraryValue) -> Result<()> {
     let mut output: Option<String> = None;
     let mut ref_output: Option<String> = None;
 
+    let input = data.to_string();
+    let brace_count = input.chars().filter(|&c| c == '{').count();
+    // Higher numbers of braces tends to cause a stack overflow (more braces
+    // use more stack space).
+    if brace_count > 350 {
+        return Ok(());
+    }
+
     rt.context().with(|cx| {
         let globals = cx.globals();
-        globals.set("INPUT", JSString::from_str(cx.clone(), &data.to_string())?)?;
+
+        globals.set("INPUT", JSString::from_str(cx.clone(), &input)?)?;
 
         let result: Result<(), _> = cx.eval(JSON_PROGRAM);
 
@@ -54,7 +69,7 @@ fn exec(data: &ArbitraryValue) -> Result<()> {
 
     ref_rt.context().with(|cx| {
         let globals = cx.globals();
-        globals.set("INPUT", JSString::from_str(cx.clone(), &data.to_string())?)?;
+        globals.set("INPUT", JSString::from_str(cx.clone(), &input)?)?;
 
         let result: Result<(), _> = cx.eval(JSON_PROGRAM);
 
@@ -71,4 +86,14 @@ fn exec(data: &ArbitraryValue) -> Result<()> {
     assert_eq!(output, ref_output);
 
     Ok(())
+}
+
+fn setup_config(config: &mut Config) {
+    config
+        // https://github.com/bellard/quickjs/blob/6e2e68fd0896957f92eb6c242a2e048c1ef3cae0/quickjs.c#L1644
+        .gc_threshold(256 * 1024)
+        // https://github.com/bellard/quickjs/blob/6e2e68fd0896957f92eb6c242a2e048c1ef3cae0/fuzz/fuzz_common.c#L33
+        .memory_limit(0x4000000)
+        // https://github.com/bellard/quickjs/blob/6e2e68fd0896957f92eb6c242a2e048c1ef3cae0/fuzz/fuzz_common.c#L35
+        .max_stack_size(0x10000);
 }